Adjust some last feature (audited) code and add feature : activity logger with some changes:

Migrate H2 to Local PSQL DB ( Due to inconsistent syntax and type
Fix Spring load new context each controller test
Others

+++ src/main/resources/db/migration/V22__create_table_memo_views.sql
@@ -5,7 +5,7 @@ CREATE TABLE memo_views (
  
     created_at TIMESTAMP NOT NULL,
     updated_at TIMESTAMP NOT NULL,
-    version INTEGER DEFAULT 0 NOT NULL,
+    version BIGINT DEFAULT 0 NOT NULL,
 
     CONSTRAINT uk_memo_employee UNIQUE (memo_id, employee_id),
