+++ src/main/resources/db/migration/V30__create_table_supply_items.sql
@@ -3,10 +3,10 @@ CREATE TABLE supply_items (
     supply_id UUID NOT NULL,
     product_id VARCHAR(55) NOT NULL,
     price BIGINT DEFAULT 0 NOT NULL,
-    quantity INTEGER DEFAULT 0 NOT NULL,
+    quantity BIGINT DEFAULT 0 NOT NULL,
     created_at TIMESTAMP NOT NULL,
     updated_at TIMESTAMP NOT NULL,
-    version INTEGER DEFAULT 0 NOT NULL,
+    version BIGINT DEFAULT 0 NOT NULL,
 
     CONSTRAINT fk_supply_items FOREIGN KEY (supply_id)
         REFERENCES supplies(id)
